refactor: extract ToolSettings list controller, row binder, and layout signature
Extract Class per Fowler: separate foldout/section shell from ListView ownership,
row VisualElement binding, and the pure rebuild-signature algorithm. Pin the
signature format with characterization tests so list rebuild decisions stay stable.
