+++ Assets/Tests/Editor/ToolSettingsSectionLayoutSignatureTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c25e44ffdb1e4c2f928333fdfd7079cd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
