+++ Packages/src/Editor/Presentation/UIToolkit/Components/ToolSettingsSectionLayoutSignature.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6b37cc27791d4505a312c78a631066e1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
